+++ tests/sources/test_web_service.py
@@ -9,9 +9,10 @@ import pytest
 import vcr
 
 from idf_component_tools.constants import IDF_COMPONENT_STORAGE_URL
-from idf_component_tools.errors import FetchingError, UserHint
+from idf_component_tools.errors import FetchingError
 from idf_component_tools.hash_tools import hash_dir
 from idf_component_tools.manifest import ComponentVersion, SolvedComponent
+from idf_component_tools.messages import UserHint
 from idf_component_tools.sources import WebServiceSource
 from idf_component_tools.sources.web_service import IDF_COMPONENT_REGISTRY_API_URL, download_archive