+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fix typo in gitlab CI user agent
+- Fix issue with creating a profile in `idf_component_manager.yml` if it does not exist, when executing `compote registry login` command with `--service-profile` specified
 
 ## [1.4.0] - 2023-09-15